ml-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
rustworkx>=0.14.0
networkx>=3.0
scipy>=1.10.0
python-louvain>=0.16
pytest>=7.0.0
httpx>=0.25.0
pydantic>=2.0.0
loguru>=0.7.0
ruff>=0.1.0
mypy>=1.0.0
black>=23.0.0
bandit>=1.7.0
